+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - `rtcontrol`: Fix `--prio`
 - `rtcontrol`: Allow spaces in regexes (e.g. `rtcontrol "message=/(not
   |un)registered/"` will work as expected now)
+- `pyrotorque`: For `QueueManager`, allow using old configuration
+  setting `sort_fields`
+- Properly handle spaces in comma-separated field lists
 
 ## [2.5.1] - 2022-12-22